[WIP] Separate static binary for chrootarchive
