+++ cmd/docker-chrootarchive/cve-2019-14271.go
@@ -1,0 +1,16 @@
+package main
+
+import (
+	"net"
+	"os/user"
+)
+
+func init() {
+	// Mitigate the root cause of CVE-2019-14721 in case all compile-time
+	// and runtime guards against dynamically linking docker-chrootarchive
+	// have been circumvented. Initialize nss libraries in Glibc so that the
+	// dynamic libraries are loaded in the host environment not in the
+	// chroot from untrusted files.
+	_, _ = user.Lookup("docker")
+	_, _ = net.LookupHost("localhost")
+}
