+++ cmd/docker-chrootarchive/notcgo.go
@@ -1,0 +1,6 @@
+//go:build !cgo
+// +build !cgo
+
+package main
+
+var thisPackageMustNotBeBuiltWithCgo struct{}
