+++ pkg/chrootarchive/archive_unix.go
@@ -6,57 +6,15 @@ package chrootarchive // import "github.com/docker/docker/pkg/chrootarchive"
 import (
 	"bytes"
 	"encoding/json"
-	"flag"
 	"fmt"
 	"io"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 
 	"github.com/docker/docker/pkg/archive"
-	"github.com/docker/docker/pkg/reexec"
 	"github.com/pkg/errors"
 )
-
-// untar is the entry-point for docker-untar on re-exec. This is not used on
-// Windows as it does not support chroot, hence no point sandboxing through
-// chroot and rexec.
-func untar() {
-	runtime.LockOSThread()
-	flag.Parse()
-
-	var options archive.TarOptions
-
-	// read the options from the pipe "ExtraFiles"
-	if err := json.NewDecoder(os.NewFile(3, "options")).Decode(&options); err != nil {
-		fatal(err)
-	}
-
-	dst := flag.Arg(0)
-	var root string
-	if len(flag.Args()) > 1 {
-		root = flag.Arg(1)
-	}
-
-	if root == "" {
-		root = dst
-	}
-
-	if err := chroot(root); err != nil {
-		fatal(err)
-	}
-
-	if err := archive.Unpack(os.Stdin, dst, &options); err != nil {
-		fatal(err)
-	}
-	// fully consume stdin in case it is zero padded
-	if _, err := flush(os.Stdin); err != nil {
-		fatal(err)
-	}
-
-	os.Exit(0)
-}
 
 func invokeUnpack(decompressedArchive io.Reader, dest string, options *archive.TarOptions, root string) error {
 	if root == "" {
@@ -87,7 +45,7 @@ func invokeUnpack(decompressedArchive io.Reader, dest string, options *archive.T
 		dest = relDest
 	}
 
-	cmd := reexec.Command("docker-untar", dest, root)
+	cmd := command("docker-untar", dest, root)
 	cmd.Stdin = decompressedArchive
 
 	cmd.ExtraFiles = append(cmd.ExtraFiles, r)
@@ -118,42 +76,6 @@ func invokeUnpack(decompressedArchive io.Reader, dest string, options *archive.T
 	return nil
 }
 
-func tar() {
-	runtime.LockOSThread()
-	flag.Parse()
-
-	src := flag.Arg(0)
-	var root string
-	if len(flag.Args()) > 1 {
-		root = flag.Arg(1)
-	}
-
-	if root == "" {
-		root = src
-	}
-
-	if err := realChroot(root); err != nil {
-		fatal(err)
-	}
-
-	var options archive.TarOptions
-	if err := json.NewDecoder(os.Stdin).Decode(&options); err != nil {
-		fatal(err)
-	}
-
-	rdr, err := archive.TarWithOptions(src, &options)
-	if err != nil {
-		fatal(err)
-	}
-	defer rdr.Close()
-
-	if _, err := io.Copy(os.Stdout, rdr); err != nil {
-		fatal(err)
-	}
-
-	os.Exit(0)
-}
-
 func invokePack(srcPath string, options *archive.TarOptions, root string) (io.ReadCloser, error) {
 	if root == "" {
 		return nil, errors.New("root path must not be empty")
@@ -175,7 +97,7 @@ func invokePack(srcPath string, options *archive.TarOptions, root string) (io.Re
 		relSrc += "/"
 	}
 
-	cmd := reexec.Command("docker-tar", relSrc, root)
+	cmd := command("docker-tar", relSrc, root)
 
 	errBuff := bytes.NewBuffer(nil)
 	cmd.Stderr = errBuff
