--- pkg/chrootarchive/init_windows.go
@@ -1,4 +1,0 @@
-package chrootarchive // import "github.com/docker/docker/pkg/chrootarchive"
-
-func init() {
-}
