+++ pkg/chrootarchive/sysprocattr_other.go
@@ -1,0 +1,10 @@
+//go:build !linux
+// +build !linux
+
+package chrootarchive // import "github.com/docker/docker/pkg/chrootarchive"
+
+import "syscall"
+
+func sysProcAttr() *syscall.SysProcAttr {
+	return nil
+}
